requirements.txt
redis
pyArango
python-dotenv==0.21.0